+++ GrpcGreeter/user.cs
@@ -2,9 +2,9 @@ namespace GrpcGreeter
 {
     public class User
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public string name { get; set; }
-        public int balance { get; set; }
+        public long balance { get; set; }
         public override string ToString()
         {
             return ($"id:{id} , {name} -- {balance}");
